Declare 'int' data type for post meta.

+++ plugins/tours/src/views/tour-title.php
@@ -8,7 +8,7 @@ namespace spiralWebDb\CornerstoneTours;
 ?>
 <h2 class="entry-title tour-title" itemprop="headline">
 	<?php echo "Tour " . esc_html( $menu_order ) . " | " .
-               esc_html( get_post_meta( $tour_id, 'tour_year', true ) ) .
+               esc_html( (int) get_post_meta( $tour_id, 'tour_year', true ) ) .
                " | " . get_the_title();
 	?>
 </h2>
